api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1

# Google Cloud
google-cloud-aiplatform==1.70.0
google-cloud-firestore==2.17.0
google-auth==2.36.0

# Optional SQL (Pro-block)
SQLAlchemy==2.0.36
pg8000==1.31.2
